Deduplicate Java/Kotlin compile scripts:
- Symlink to common DetectMain.java source.
- Fix left-over "Java" naming in Kotlin script.
